fix: rename adapter
